final v1

+++ app/controllers/api/v1/items_controller.rb
@@ -47,15 +47,9 @@ class Api::V1::ItemsController < ApplicationController
         crm: "cream"
     }
 
-    def practices
-        begin
-            @respo = ActiveRecord::Base.connection.execute("select * from custom_practices")
-            render_json(200, @respo.count.to_s+" record/s found!",@respo)
-        rescue => e
-            @error = e.message
-           render_json(200, @error,[]) 
-        end
-    end
+
+
+   
 
     def mapped_items
         begin
@@ -76,7 +70,11 @@ class Api::V1::ItemsController < ApplicationController
         begin
             practice_id = params[:practice_id]
             if practice_id
-                @respo = ActiveRecord::Base.connection.execute("SELECT d.id as def_id,d.item_name as def_item_name, d.custom_id as def_custom_id, d.categories as def_categories, d.param_type as def_param_type, d.param_value as def_param_value, d.name as def_name, d.remaining_name as def_remaining_name, d.second_name as def_second_name,f.id as p_id,f.item_name as p_item_name, f.custom_id as p_custom_id, f.categories as p_categories, f.param_type as p_param_type, f.param_value as p_param_value, f.name as p_name, f.remaining_name as p_remaining_name, f.second_name as p_second_name, case when m.id is not null then 'mapped' else '' end as status FROM public.formatted_items as f inner join public.formatted_items d on d.item_name = f.item_name and d.custom_practice_id != #{practice_id} left outer join mapped_items m on f.id = m.formatted_item_id where m.id is null and f.custom_practice_id = #{practice_id} ")
+                @respo = ActiveRecord::Base.connection.execute("SELECT distinct d.id as def_id,d.item_name as def_item_name, d.custom_id as def_custom_id, d.categories as def_categories, d.param_type as def_param_type, d.param_value as def_param_value, d.name as def_name, d.remaining_name as def_remaining_name, d.second_name as def_second_name,f.id as p_id,f.item_name as p_item_name, f.custom_id as p_custom_id, f.categories as p_categories, f.param_type as p_param_type, f.param_value as p_param_value, f.name as p_name, f.remaining_name as p_remaining_name, f.second_name as p_second_name, case when m.id is not null then 'mapped' else '' end as status,m.id as mapped_item FROM public.formatted_items as f inner join public.formatted_items d on d.item_name = f.item_name and d.custom_practice_id != #{practice_id} left outer join mapped_items m on f.id = m.formatted_item_id where m.id is null and f.custom_practice_id = #{practice_id}")
+
+
+            
+
                 render_json(200, @respo.count.to_s+" record/s found!",@respo)
             else
                 render_json(400, "Practice Id is required!",[])
@@ -91,14 +89,15 @@ class Api::V1::ItemsController < ApplicationController
         begin
             practice_id = params[:practice_id]
             if practice_id
-                @respo = ActiveRecord::Base.connection.execute("SELECT d.id as def_id,d.item_name as def_item_name, d.custom_id as def_custom_id, d.categories as def_categories, d.param_type as def_param_type, d.param_value as def_param_value, d.name as def_name, d.remaining_name as def_remaining_name, d.second_name as def_second_name,f.id as p_id,f.item_name as p_item_name, f.custom_id as p_custom_id, f.categories as p_categories, f.param_type as p_param_type, f.param_value as p_param_value, f.name as p_name, f.remaining_name as p_remaining_name, f.second_name as p_second_name, case when m.id is not null then 'mapped' else '' end as status  FROM public.formatted_items as f inner join public.formatted_items d on d.name = f.name and d.custom_practice_id != #{practice_id} and d.second_name = f.second_name and d.param_type = f.param_type and d.param_value = f.param_value and d.categories = f.categories left outer join mapped_items m on f.id = m.formatted_item_id where f.custom_practice_id =  #{practice_id}  and m.id is null and f.id not in (select * from (SELECT distinct f.id FROM public.default_items as d inner join public.formatted_items f on d.item_name = f.item_name where f.custom_practice_id = #{practice_id}) as a)")
+                @respo = ActiveRecord::Base.connection.execute("SELECT distinct d.id as def_id,d.item_name as def_item_name, d.custom_id as def_custom_id, d.categories as def_categories, d.param_type as def_param_type, d.param_value as def_param_value, d.name as def_name, d.remaining_name as def_remaining_name, d.second_name as def_second_name,f.id as p_id,f.item_name as p_item_name, f.custom_id as p_custom_id, f.categories as p_categories, f.param_type as p_param_type, f.param_value as p_param_value, f.name as p_name, f.remaining_name as p_remaining_name, f.second_name as p_second_name, case when m.id is not null then 'mapped' else '' end as status,m.id as mapped_item  FROM public.formatted_items as f inner join public.formatted_items d on d.name = f.name and d.custom_practice_id != #{practice_id} and d.second_name = f.second_name and d.param_type = f.param_type and d.param_value = f.param_value and d.categories = f.categories left outer join mapped_items m on f.id = m.formatted_item_id where f.custom_practice_id =  #{practice_id}  and m.id is null ")
+                # and f.id not in (select * from (SELECT distinct f.id FROM public.default_items as d inner join public.formatted_items f on d.item_name = f.item_name where f.custom_practice_id = #{practice_id}) as a)
                 render_json(200, @respo.count.to_s+" record/s found!",@respo)
             else
                 render_json(400, "Practice Id is required!",[])
             end
         rescue => e
             @error = e.message
-           render_json(200, @error,[]) 
+           render_json(500, @error,[]) 
         end
     end
 
@@ -106,7 +105,12 @@ class Api::V1::ItemsController < ApplicationController
         begin
             practice_id = params[:practice_id]
             if practice_id
-                @respo = ActiveRecord::Base.connection.execute("SELECT d.id as def_id,f.id as p_id,d.item_name as def_item_name, d.custom_id as def_custom_id, d.categories as def_categories, d.param_type as def_param_type, d.param_value as def_param_value, d.name as def_name, d.remaining_name as def_remaining_name, d.second_name as def_second_name,f.item_name as p_item_name, f.custom_id as p_custom_id, f.categories as p_categories, f.param_type as p_param_type, f.param_value as p_param_value, f.name as p_name, f.remaining_name as p_remaining_name, f.second_name as p_second_name FROM public.formatted_items as f inner join public.formatted_items d on d.name = f.name and d.custom_practice_id != #{practice_id} and d.param_type = f.param_type and d.param_value = f.param_value and d.categories = f.categories left outer join mapped_items m on f.id = m.formatted_item_id where m.id is null and f.custom_practice_id = #{practice_id} and f.param_type != '' and f.id not in (select * from (SELECT distinct f.id FROM public.default_items as d inner join public.formatted_items f on d.item_name = f.item_name where f.custom_practice_id = #{practice_id} union all SELECT distinct f.id FROM public.default_items as d inner join public.formatted_items f on d.name = f.name and d.second_name = f.second_name and d.param_type = f.param_type and d.param_value = f.param_value and d.categories = f.categories where f.custom_practice_id =  #{practice_id}) as a)")
+
+                # @respo = FormattedItem.joins("inner join public.formatted_items d on d.name = formatted_items.name and d.custom_practice_id != #{practice_id} and d.param_type = formatted_items.param_type and d.param_value = formatted_items.param_value and d.categories = formatted_items.categories left outer join mapped_items m on formatted_items.id = m.formatted_item_id where m.id is null and formatted_items.custom_practice_id = #{practice_id} and formatted_items.param_type != '' and formatted_items.id not in (select * from (SELECT distinct formatted_items.id FROM public.default_items as d inner join public.formatted_items f on d.item_name = formatted_items.item_name where formatted_items.custom_practice_id = #{practice_id} union all SELECT distinct formatted_items.id FROM public.default_items as d inner join public.formatted_items f on d.name = formatted_items.name and d.second_name = formatted_items.second_name and d.param_type = formatted_items.param_type and d.param_value = formatted_items.param_value and d.categories = formatted_items.categories where formatted_items.custom_practice_id =  #{practice_id}) as a)")
+
+                @respo = ActiveRecord::Base.connection.execute("SELECT distinct d.id as def_id,f.id as p_id,d.item_name as def_item_name, d.custom_id as def_custom_id, d.categories as def_categories, d.param_type as def_param_type, d.param_value as def_param_value, d.name as def_name, d.remaining_name as def_remaining_name, d.second_name as def_second_name,f.item_name as p_item_name, f.custom_id as p_custom_id, f.categories as p_categories, f.param_type as p_param_type, f.param_value as p_param_value, f.name as p_name, f.remaining_name as p_remaining_name, f.second_name as p_second_name, case when m.id is not null then 'mapped' else '' end as status,m.id as mapped_item  FROM public.formatted_items as f inner join public.formatted_items d on d.name = f.name and d.custom_practice_id != #{practice_id} and d.param_type = f.param_type and d.param_value = f.param_value and d.categories = f.categories left outer join mapped_items m on f.id = m.formatted_item_id where f.custom_practice_id = #{practice_id} and f.param_type != ''")
+
+                #and f.id not in (select * from (SELECT distinct f.id FROM public.default_items as d inner join public.formatted_items f on d.item_name = f.item_name where f.custom_practice_id = #{practice_id} union all SELECT distinct f.id FROM public.default_items as d inner join public.formatted_items f on d.name = f.name and d.second_name = f.second_name and d.param_type = f.param_type and d.param_value = f.param_value and d.categories = f.categories where f.custom_practice_id =  #{practice_id}) as a)
                 render_json(200, @respo.count.to_s+" record/s found!",@respo)
             else
                 render_json(400, "Practice Id is required!",[])
@@ -121,7 +125,9 @@ class Api::V1::ItemsController < ApplicationController
         begin
             practice_id = params[:practice_id]
             if practice_id
-                @respo = ActiveRecord::Base.connection.execute("select * from formatted_items f  left outer join mapped_items m on f.id = m.formatted_item_id  where m.id is null and f.custom_practice_id =  #{practice_id} and f.id not in (select * from (SELECT distinct f.id FROM public.default_items as d inner join public.formatted_items f on d.item_name = f.item_name where f.custom_practice_id = #{practice_id} union all SELECT distinct f.id FROM public.default_items as d inner join public.formatted_items f on d.name = f.name and d.second_name = f.second_name and d.param_type = f.param_type and d.param_value = f.param_value and d.categories = f.categories where f.custom_practice_id =  #{practice_id} union all SELECT distinct f.id FROM public.default_items as d inner join public.formatted_items f on d.name = f.name and d.param_type = f.param_type and d.param_value = f.param_value and d.categories = f.categories where f.custom_practice_id = #{practice_id} and f.param_type != '' ) as a)")
+
+                @respo = ActiveRecord::Base.connection.execute("select distinct f.*, case when m.id is not null then 'mapped' else '' end as status,m.id as mapped_item   from formatted_items f  left outer join mapped_items m on f.id = m.formatted_item_id  where m.id is null and f.custom_practice_id =  #{practice_id}")
+
                 render_json(200, @respo.count.to_s+" record/s found!",@respo)
             else
                 render_json(400, "Practice Id is required!",[])
@@ -156,6 +162,7 @@ class Api::V1::ItemsController < ApplicationController
     def create
         data = params["data"]
         type = params["type"] == "0" 
+        delete_previous = params["delete"]
         practice_id = params["practice_id"]
         begin
             if type
@@ -167,7 +174,13 @@ class Api::V1::ItemsController < ApplicationController
                     # data = JSON.load file
                     @response = manipulate(data,1)
                     @filter = @response.select { |row| row["item_name"] != '' }
-                    @inserted = DefaultItem.insert_all(@filter)
+                    
+                    if delete_previous 
+                        ActiveRecord::Base.connection.execute("truncate default_items")
+                        @inserted = DefaultItem.insert_all(@filter)
+                    else 
+                        @inserted = DefaultItem.insert_all(@filter)
+                    end
                     render_json(200, @inserted.length.to_s+" record/s found!",@inserted)
                     # @respo = {inserted_objects: @filter.count, response: @filter}
                 end
@@ -180,8 +193,13 @@ class Api::V1::ItemsController < ApplicationController
                 else
                     @response = manipulate(data,practice_id)
                     @filter = @response.select { |row| row["item_name"] != '' }
-                    ActiveRecord::Base.connection.execute("delete from formatted_items where custom_practice_id = #{practice_id}")
-                    @inserted = FormattedItem.insert_all(@filter)
+                    if delete_previous 
+                        ActiveRecord::Base.connection.execute("delete from formatted_items where custom_practice_id = #{practice_id}")
+                        @inserted = FormattedItem.insert_all(@filter)
+                    else
+                        @inserted = FormattedItem.insert_all(@filter)
+                    end
+                    
                     render_json(200, @inserted.length.to_s+" record/s found!",@inserted)
                 end
             end
